# Conflicts:
#	src/main/java/com/sparta/spartdelivery/controller/CartController.java
#	src/main/java/com/sparta/spartdelivery/controller/OrderController.java
#	src/main/java/com/sparta/spartdelivery/controller/StoreController.java
#	src/main/java/com/sparta/spartdelivery/external/security/WebSecurityConfig.java
#	src/main/java/com/sparta/spartdelivery/service/CartService.java
#	src/main/java/com/sparta/spartdelivery/service/OrderService.java
#	src/main/java/com/sparta/spartdelivery/service/ReviewService.java
#	src/main/resources/static/js/cart.js
#	src/main/resources/static/js/store-detail.js
#	src/main/resources/templates/cart.html

+++ src/main/java/com/sparta/spartdelivery/repository/MenuRepository.java
@@ -4,10 +4,10 @@ import com.sparta.spartdelivery.entity.Menu;
 import com.sparta.spartdelivery.entity.Store;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
 import java.util.List;
 
 @Repository
-public interface MenuRepository extends JpaRepository<Menu, Long> {
+public interface MenuRepository extends JpaRepository<Menu, Integer> {
     List<Menu> findByStore(Store store);
-
 }
