+++ src/main/java/com/sparta/spartdelivery/repository/ReviewRepository.java
@@ -1,12 +1,11 @@
 package com.sparta.spartdelivery.repository;
 
 import com.sparta.spartdelivery.entity.Review;
-import com.sparta.spartdelivery.entity.Store;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ReviewRepository extends JpaRepository<Review, Long> {
+public interface ReviewRepository extends JpaRepository<Review, Integer> {
 
     List<Review> findByStore_storeId(Integer storeId);
